# Use read lock for DeltaFIFO#HasSynced