**What type of PR is this?**
/kind cleanup

**What this PR does / why we need it**:
DeltaFIFO#HasSynced only performs read operation.

This PR changes the lock type to read lock.


```release-note
NONE
```