@tedyu 2019-07-14
/test pull-kubernetes-bazel-test