@tedyu 2019-08-24
The following result was obtained by transforming TestFIFO_HasSynced into BenchmarkFIFO_HasSynced
```
+       t.ReportAllocs()
+       t.ResetTimer()
+       for n := 1; n <= t.N; n++ {
+               for i, test := range tests {
```
Without read lock:

BenchmarkFIFO_HasSynced-12             	  500000	      3194 ns/op	    3072 B/op	      40 allocs/op
BenchmarkFIFO_HasSynced-12             	  500000	      3205 ns/op	    3072 B/op	      40 allocs/op
BenchmarkFIFO_HasSynced-12             	  500000	      3226 ns/op	    3072 B/op	      40 allocs/op

With read lock:

BenchmarkFIFO_HasSynced-12             	  500000	      3135 ns/op	    3072 B/op	      40 allocs/op
BenchmarkFIFO_HasSynced-12             	  500000	      3132 ns/op	    3072 B/op	      40 allocs/op